Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/dom/nodes/Document-constructor-svg-expected.txt

CONSOLE MESSAGE: line 87: TypeError: null is not an object (evaluating 'document.body.appendChild')
CONSOLE MESSAGE: line 2237: TypeError: undefined is not an object (evaluating 'properties.output_document')
#PID UNRESPONSIVE - com.apple.WebKit.WebContent.Development (pid 52955)
FAIL: Timed out waiting for notifyDone to be called

#EOF
#EOF
